Refactor files into packages according to Golang standards
