+++ cmd/news/Dockerfile
@@ -1,0 +1,5 @@
+FROM alpine
+RUN apk add --no-cache ca-certificates
+COPY news /news
+EXPOSE 443
+ENTRYPOINT ["/news"]
