+++ pkg/handlers/context.go
@@ -1,8 +1,8 @@
 package handlers
 
 import (
-	"github.com/2charm/spectrum-api/gateway/models/sessions"
-	"github.com/2charm/spectrum-api/gateway/models/users"
+	"github.com/2charm/spectrum-api/pkg/sessions"
+	"github.com/2charm/spectrum-api/pkg/users"
 )
 
 //HandlerContext represents a receiver for handlers to utilize and gain context
