+++ pkg/handlers/cors.go
