+++ pkg/users/mysqlstore.go
@@ -29,7 +29,7 @@ func NewMySQLStore(db *sql.DB) *MySQLStore {
 //GetByID returns the User with the given ID
 func (mss *MySQLStore) GetByID(id int64) (*User, error) {
 	user := &User{}
-	row := mss.Client.QueryRow("select * from users where id=?", id)
+	row := mss.Client.QueryRow("select * from users where user_id=?", id)
 	if err := row.Scan(&user.ID, &user.Email, &user.PassHash, &user.UserName,
 		&user.FirstName, &user.LastName); err != nil {
 		return nil, err
@@ -80,7 +80,7 @@ func (mss *MySQLStore) Insert(user *User) (*User, error) {
 //Update applies UserUpdates to the given user ID
 //and returns the newly-updated user
 func (mss *MySQLStore) Update(id int64, updates *Updates) (*User, error) {
-	insq := "update users set first_name=?, last_name=? where id=?"
+	insq := "update users set first_name=?, last_name=? where user_id=?"
 	_, err := mss.Client.Exec(insq, updates.FirstName, updates.LastName, id)
 	if err != nil {
 		return nil, err
@@ -90,7 +90,7 @@ func (mss *MySQLStore) Update(id int64, updates *Updates) (*User, error) {
 
 //Delete deletes the user with the given ID
 func (mss *MySQLStore) Delete(id int64) error {
-	insq := "delete from users where id=?"
+	insq := "delete from users where user_id=?"
 	_, err := mss.Client.Exec(insq, id)
 	if err != nil {
 		return ErrUserNotFound
