+++ scripts/deploy.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-./build.sh
+../build/build.sh
 docker push 2charm/gateway
 docker push 2charm/sql
 scp -i "~/downloads/privatekey.pem" ../news_api.key ec2-user@ec2-3-18-100-210.us-east-2.compute.amazonaws.com:~/
